pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Aditya Jha <4adityajha@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
tenacity = "^8.3.0"
odmantic = "^1.0.2"
motor = "^3.4.0"
pymongo = "^4.7.2"
pydantic-settings = "^2.2.1"
passlib = "^1.7.4"
argon2-cffi = "^23.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
